+++ tests/json_util_test.py
@@ -8,11 +8,12 @@ from unittest import mock
 
 import pytest
 import test_util
+from cryptography import x509
 
 from josepy import errors, interfaces, util
 
-CERT = test_util.load_comparable_cert("cert.pem")
-CSR = test_util.load_comparable_csr("csr.pem")
+CERT = test_util.load_cert("cert.pem")
+CSR = test_util.load_csr("csr.pem")
 
 
 class FieldTest(unittest.TestCase):
@@ -327,7 +328,7 @@ class DeEncodersTest(unittest.TestCase):
         from josepy.json_util import decode_cert
 
         cert = decode_cert(self.b64_cert)
-        assert isinstance(cert, util.ComparableX509)
+        assert isinstance(cert, x509.Certificate)
         assert cert == CERT
         with pytest.raises(errors.DeserializationError):
             decode_cert("")
@@ -341,7 +342,7 @@ class DeEncodersTest(unittest.TestCase):
         from josepy.json_util import decode_csr
 
         csr = decode_csr(self.b64_csr)
-        assert isinstance(csr, util.ComparableX509)
+        assert isinstance(csr, x509.CertificateSigningRequest)
         assert csr == CSR
         with pytest.raises(errors.DeserializationError):
             decode_csr("")